Use a cuda graph for each input shape
Since llama has different stages: compute input, and generate
output, etc. The input shape, input length in particular, might
be different.

Cuda graph requires that the input shape to be consistent. So we
use the shape as key, create a cuda graph for each input shape.

NOTE: when testing w/ 2 adapters, there is no problem. When
	testing w/ 64 adapters, there was a abort + core dump, but
	no error message. Not sure why, still need to test.